+++ packages/plugin-packages/mashroom-http-proxy/src/proxy/mashroom-bootstrap-services.ts
@@ -1,6 +1,7 @@
 
 import MashroomHttpProxyService from './MashroomHttpProxyService';
 import {setPoolConfig} from '../connection_pool';
+import context from '../context/global_context';
 import {startExportPoolMetrics, stopExportPoolMetrics} from '../metrics/connection_pool_metrics';
 
 import type {MashroomServicesPluginBootstrapFunction} from '@mashroom/mashroom/type-definitions';
@@ -14,7 +15,7 @@ const bootstrap: MashroomServicesPluginBootstrapFunction = async (pluginName, pl
     })
 
     const pluginContext = pluginContextHolder.getPluginContext();
-    const service = new MashroomHttpProxyService(forwardMethods, forwardHeaders, socketTimeoutMs, pluginContext.loggerFactory);
+    const service = new MashroomHttpProxyService(forwardMethods, forwardHeaders, socketTimeoutMs, context.pluginRegistry, pluginContext.loggerFactory);
 
     startExportPoolMetrics(pluginContextHolder);
     pluginContext.services.core.pluginService.onUnloadOnce(pluginName, () => {