+++ packages/plugin-packages/mashroom-http-proxy/type-definitions/index.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type {ExpressRequest, ExpressResponse} from '@mashroom/mashroom/type-definitions';
+import type {ExpressRequest, ExpressResponse, MashroomPluginConfig, MashroomPluginContextHolder} from '@mashroom/mashroom/type-definitions';
 
 export type HttpHeaders = {
     [string]: string,
@@ -16,3 +16,30 @@ export interface MashroomHttpProxyService {
     forward(req: ExpressRequest, res: ExpressResponse, targetUri: string, additionalHeaders?: HttpHeaders): Promise<void>;
 
 }
+
+export type MashroomHttpProxyInterceptorResult = {
+    addHeaders?: HttpHeaders;
+    removeHeaders?: Array<string>;
+    rewrittenTargetUri?: string;
+    reject?: boolean;
+    rejectStatusCode?: number;
+    rejectReason?: string;
+}
+
+interface MashroomHttpProxyInterceptor {
+
+    /**
+     * Intercept a call to given targetUri.
+     * The additionalHeaders contain headers already added by the caller and other MashroomHttpProxyInterceptor plugins.
+     *
+     * Return null or undefined if you don't want to interfere with a call.
+     */
+    intercept(req: ExpressRequest, additionalHeaders: HttpHeaders, targetUri: string): Promise<?MashroomHttpProxyInterceptorResult>;
+}
+
+/*
+ * Bootstrap method definition for http-proxy-interceptor plugins
+ */
+export type MashroomHttpProxyInterceptorPluginBootstrapFunction = (pluginName: string, pluginConfig: MashroomPluginConfig, contextHolder: MashroomPluginContextHolder) => Promise<MashroomHttpProxyInterceptor>;
+
+