+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## [unreleased]
 
+ * Http Proxy: Added a new plugin type *http-proxy-interceptor* which allows it to rewrite target URIs and headers.
+   This can be used to add security headers to backend calls.
+   **BREAKING CHANGE**: The *getApiSecurityHeaders()* method in the security provider interfaces has been removed since
+   the http-proxy-interceptor is the more generic approach to solve the same problem.
  * LDAP Security Provider: Use the LDAP attributes *displayName* or *givenName* + *sn* as displayName instead of *cn*
  * Portal: The *mashroom-portal-demo-alternative-theme* module uses now *express-react-views* and *TypeScript* to demonstrate
    a type save theme template